Fix: cambios para una estructura más simple
