--- schemas/__init__.py
@@ -1,1 +1,0 @@
-# Schemas de validación para requests y responses
