--- utils/Pipfile
@@ -1,15 +1,0 @@
-[[source]]
-url = "https://pypi.org/simple"
-verify_ssl = true
-name = "pypi"
-
-[packages]
-fastapi = "==0.104.1"
-pydantic = "==2.5.0"
-python-multipart = "==0.0.6"
-uvicorn = {extras = ["standard"], version = "==0.24.0"}
-
-[dev-packages]
-
-[requires]
-python_version = "3.13"
